fix(storybook): keep facet router thin and container-responsive
